Remove unused server functions and storage functions from blocktx
