+++ blocktx/blocktx_api/blocktx_api.proto
@@ -17,24 +17,11 @@ service BlockTxAPI {
   // GetTransactionMerklePath returns the merkle path of a transaction.
   rpc GetTransactionMerklePath(Transaction) returns (MerklePath) {}
 
-  // GetBlockTransactions returns a list of transaction hashes for a given block.
-  rpc GetBlockTransactions (Block) returns (Transactions) {}
-
   // GetTransactionBlocks returns a list of block hashes (excluding orphaned) for a given transaction hash.
   rpc GetTransactionBlocks (Transactions) returns (TransactionBlocks) {}
 
-  // GetBlock returns the non-orphaned block for a given block hash.
-  rpc GetBlock (Hash) returns (Block) {}
-
   // GetBlockForHeight returns the non-orphaned block for a given block height.
   rpc GetBlockForHeight (Height) returns (Block) {}
-
-  // GetLastProcessedBlock returns the last processed block.
-  rpc GetLastProcessedBlock (google.protobuf.Empty) returns (Block) {}
-
-  // GetMyTransactionsForBlock returns a list of transaction hashes for a given block that were registered by this API.
-  rpc GetMinedTransactionsForBlock(BlockAndSource) returns (MinedTransactions) {}
-
 }
 
 // swagger:model HealthResponse
