--- blocktx/store/sql/orphan_height.go
@@ -1,38 +1,0 @@
-package sql
-
-import (
-	"context"
-	"database/sql"
-
-	"github.com/ordishs/gocore"
-)
-
-func (s *SQL) OrphanHeight(ctx context.Context, height uint64) error {
-	start := gocore.CurrentNanos()
-	defer func() {
-		gocore.NewStat("blocktx").NewStat("OrphanHeight").AddTime(start)
-	}()
-
-	return setOrphanedForHeight(ctx, s.db, height, true)
-}
-
-func (s *SQL) SetOrphanHeight(ctx context.Context, height uint64, orphaned bool) error {
-	return setOrphanedForHeight(ctx, s.db, height, orphaned)
-}
-
-func setOrphanedForHeight(ctx context.Context, db *sql.DB, height uint64, orphaned bool) error {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	q := `
-		UPDATE blocks
-		SET orphanedyn = $1
-		WHERE height = $2
-	`
-
-	if _, err := db.ExecContext(ctx, q, orphaned, height); err != nil {
-		return err
-	}
-
-	return nil
-}
